Add first version of the script
This first version of the script creates hard links, not symbolic links.
The links are stored in ~/.fonts, which is a deprecated location which
will be removed in the future, according to a comment in the Fontconfig
configuration. And it differentiates services by adding a custom extension
to the name of each link file instead of using a separate subfolder.

I am adding successive versions of the script for historical purpose.
I have corrected each of these shortcomings in a later version.

This script is similar to the version I used in February 2020.