Create symbolic links instead of hard links
Hard links are low-level pointers on the hidden fonts created by the font
subscription service. Using hard links prevents the files from being deleted
when they are removed from the cache of the font subscription service. They
are preserved because the hard links are still pointing on them. They remain
on the system until the script is called again to refresh the list.

This script is similar to the version I used in March 2021.